% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwc-ta.R
\name{add_macd}
\alias{add_macd}
\title{Add MACD in a separate pane by default}
\usage{
add_macd(
  chart,
  nFast = 12,
  nSlow = 26,
  nSig = 9,
  column = NULL,
  pane = "macd",
  pane_height = 1,
  ...
)
}
\arguments{
\item{chart}{A \code{lightweightchartR} chart object.}

\item{nFast, nSlow, nSig}{MACD periods.}

\item{column}{Source column. Defaults to the close column.}

\item{pane}{Target pane.}

\item{pane_height}{Relative pane height for the RSI pane.}

\item{...}{Reserved for future extensions.}
}
\value{
The updated chart object.
}
\description{
Add MACD in a separate pane by default
}
