% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwc-chart.R
\name{lightweightchartR-shiny}
\alias{lightweightchartR-shiny}
\alias{lightweightchartROutput}
\alias{renderLightweightchartR}
\title{Shiny bindings for lightweightchartR}
\usage{
lightweightchartROutput(outputId, width = "100\%", height = "480px")

renderLightweightchartR(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{CSS dimensions for the widget container.}

\item{expr}{An expression that generates a chart.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} already quoted?}
}
\description{
Shiny bindings for lightweightchartR
}
