% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightweightchartR-package.R
\docType{package}
\name{lightweightchartR-package}
\alias{lightweightchartR}
\alias{lightweightchartR-package}
\title{lightweightchartR}
\description{
Pipe-first TradingView lightweight charts for quantmod workflows.
}
\author{
\strong{Maintainer}: Alex Lightweight \email{alex@example.com}

}
\keyword{internal}
